# docs/v2/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "AI Agent Productivity Platform API V2",
        "contact": {},
        "version": "v0.0.1"
    },
    "basePath": "/v2",
    "paths": {
        "/statistic/api/chart": {
            "post": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "API Key调用趋势 + API Key调用次数排行",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "app_observability.statistic"
                ],
                "summary": "获取API Key统计趋势+排行",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.APIKeyStatisticV2ChartReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.APIKeyStatisticV2Chart"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/statistic/api/list": {
            "post": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "按API Key+路径聚合的调用统计（分页）。sortField 可选数值指标：callCount、callFailure、streamCount、nonStreamCount、streamFailure、nonStreamFailure、failureRate、avgFirstTokenLatency、avgCosts、firstTokenLatency、costs",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "app_observability.statistic"
                ],
                "summary": "获取API Key调用统计列表",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.APIKeyStatisticV2ListReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/response.PageResult"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "list": {
                                                            "type": "array",
                                                            "items": {
                                                                "$ref": "#/definitions/response.APIKeyStatisticV2ListItem"
                                                            }
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/statistic/api/list/app": {
            "post": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "指定API Key下的应用调用统计（分页）。sortField 可选数值指标：callCount、callFailure、streamCount、nonStreamCount、streamFailure、nonStreamFailure、failureRate、avgFirstTokenLatency、avgCosts、firstTokenLatency、costs",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "app_observability.statistic"
                ],
                "summary": "获取API Key应用调用统计列表",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.APIKeyStatisticV2AppListReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/response.PageResult"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "list": {
                                                            "type": "array",
                                                            "items": {
                                                                "$ref": "#/definitions/response.APIKeyStatisticV2AppListItem"
                                                            }
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/statistic/api/list/app/export": {
            "post": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "导出API Key应用调用统计列表。sortField 可选数值指标：callCount、callFailure、streamCount、nonStreamCount、streamFailure、nonStreamFailure、failureRate、avgFirstTokenLatency、avgCosts、firstTokenLatency、costs",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "app_observability.statistic"
                ],
                "summary": "导出API Key应用调用统计列表",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.APIKeyStatisticV2AppExportReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/statistic/api/list/export": {
            "post": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "导出API Key调用统计列表。sortField 可选数值指标：callCount、callFailure、streamCount、nonStreamCount、streamFailure、nonStreamFailure、failureRate、avgFirstTokenLatency、avgCosts、firstTokenLatency、costs",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "app_observability.statistic"
                ],
                "summary": "导出API Key调用统计列表",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.APIKeyStatisticV2ExportListReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/statistic/api/list/model": {
            "post": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "指定API Key下的模型使用统计（分页）。钻取需传当前主表行的 apiKeyId/methodPath。sortField 可选数值指标：totalTokens、promptTokens、completionTokens、callCount、callFailure、failureRate、avgFirstTokenLatency、avgCosts",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "app_observability.statistic"
                ],
                "summary": "获取API Key模型使用列表",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.APIKeyStatisticV2ModelListReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/response.PageResult"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "list": {
                                                            "type": "array",
                                                            "items": {
                                                                "$ref": "#/definitions/response.APIKeyStatisticV2ModelListItem"
                                                            }
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/statistic/api/list/model/export": {
            "post": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "导出API Key模型使用列表。sortField 可选数值指标：totalTokens、promptTokens、completionTokens、callCount、callFailure、failureRate、avgFirstTokenLatency、avgCosts",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "app_observability.statistic"
                ],
                "summary": "导出API Key模型使用列表",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.APIKeyStatisticV2ModelExportReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/statistic/api/overview": {
            "post": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "10张指标卡（调用总次数、失败次数、日均调用/失败、日均流式/非流式、平均耗时、流式/非流式调用次数）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "app_observability.statistic"
                ],
                "summary": "获取API Key统计概览",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.APIKeyStatisticV2Req"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.APIKeyStatisticV2Overview"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/statistic/api/record/export": {
            "post": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "导出API Key调用明细（不支持用户排序；后端固定按调用时间倒序）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "app_observability.statistic"
                ],
                "summary": "导出API Key调用明细",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.APIKeyStatisticV2RecordExportReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/statistic/api/record/list": {
            "post": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "API Key调用明细记录（分页，不支持用户排序；后端固定按调用时间倒序）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "app_observability.statistic"
                ],
                "summary": "获取API Key调用明细列表",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.APIKeyStatisticV2RecordReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/response.PageResult"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "list": {
                                                            "type": "array",
                                                            "items": {
                                                                "$ref": "#/definitions/response.APIKeyStatisticV2RecordItem"
                                                            }
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "request.APIKeyStatisticV2ChartReq": {
            "type": "object",
            "required": [
                "endDate",
                "startDate"
            ],
            "properties": {
                "apiKeyIds": {
                    "description": "空=全部",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "endDate": {
                    "type": "string"
                },
                "limit": {
                    "description": "排行 TopN，默认 5",
                    "type": "integer"
                },
                "methodPaths": {
                    "description": "空=全部，如 POST-/v1/chat/completions",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "orgIds": {
                    "description": "空=按角色解析（见上方约定）；[\"ALL\"]=可见全部组织；[\"id\",...]=指定组织",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "startDate": {
                    "description": "yyyy-MM-dd",
                    "type": "string"
                },
                "userIds": {
                    "description": "空=按角色解析（见上方约定）；[\"ALL\"]=已解析组织下全部用户；[\"id\",...]=指定用户",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "request.APIKeyStatisticV2ListReq": {
            "type": "object",
            "required": [
                "endDate",
                "pageNo",
                "pageSize",
                "startDate"
            ],
            "properties": {
                "apiKeyIds": {
                    "description": "空=全部",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "endDate": {
                    "type": "string"
                },
                "methodPaths": {
                    "description": "空=全部，如 POST-/v1/chat/completions",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "orgIds": {
                    "description": "空=按角色解析（见上方约定）；[\"ALL\"]=可见全部组织；[\"id\",...]=指定组织",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "pageNo": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "sortField": {
                    "description": "见各列表接口 Description 中的可选值",
                    "type": "string",
                    "example": "callCount"
                },
                "sortOrder": {
                    "description": "asc|desc，默认 desc",
                    "type": "string",
                    "example": "desc"
                },
                "startDate": {
                    "description": "yyyy-MM-dd",
                    "type": "string"
                },
                "userIds": {
                    "description": "空=按角色解析（见上方约定）；[\"ALL\"]=已解析组织下全部用户；[\"id\",...]=指定用户",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "response.APIKeyStatisticV2Chart": {
            "type": "object",
            "properties": {
                "rank": {
                    "$ref": "#/definitions/response.APIKeyStatisticV2Rank"
                },
                "trend": {
                    "$ref": "#/definitions/response.APIKeyStatisticV2Trend"
                }
            }
        },
...
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "JWT": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}